AD-HOC feat (Tasks): Allow opting out
It is the authors opinion that when ansible tasks create resources or
modify state, it should be possible for the playbook to "opt out" of the
role simply by specifying that there be zero state created. An example
of this would be a role that managers linux users. The task (in psuedo
code) does:

  for all users; do
    create user;
  done;

Where there are zero users present, there are zero users created.

This commit replicates this behaviour by skipping the entire role if the
variable "letsencrypt_common_name" is undefined. This is a required
variable; it's absence means the task will fail. So, it can reasonably
be skipped given it's absence.

Design Notes:

alternative design: allow multiple certificates

  A better way to do this would be to extend the role to allow multiple
  certificate provisioning. In this way, the "bare" resources can be
  created (i.e. the Lets Encrypt signing key) but no actual certificates
  provisioned. However, the author does not have the time to complete
  this change.