AD-HOC refactor (*): Restructure provider mechanism to allow specific details per domain
Currently, there is a problem in which a single certificate cannot be
used with the DNS-01 mechanism for multiple top level domains. Multiple
top level domains using the SAN mechanism to supply the certificates is
valid[1], and thus should be supported for this role.

To accomplish, the abstraction of domains was changed slightly from a
single set of configuration with multiple SANs, to a set of domains
including configuration information that is concatenated to make the
list of SANs.

For example, instead of this:

  # Psuedo

  - domain.1.de
  - domain.2.de
  - domain.3.de

It becomes:

  # psuedo

  - name: domain.1.de
    configuration:
      zone: 1.de

This allows specifying configuration options on a per SAN basis, such as
in which zone the SAN needs to be verified. This is then implemented in
the route53 role to allow multiple TLDs for the provided certificates.

Design Notes:

alternative design: reuse the role

  It's also reasonable to reuse the role, and simply issue a different
  certificate for each domain. However, it is reasonable for multiple
  TLD SANs to be on a single certificate; it's thusly a supported case
  within the role abstraction, and does not require multiple
  invocations.

future thinking about http verification:

  This model also work quite well for HTTP verification, in which a
  number of public dirs may be used for a single certificate issuance.

code style:

  This commit also includes a number of minor code style changes.

no conditional on pause:

  The conditional on pause was removed as part of this work. The
  author considered the pause, and decided to remove it as it's
  reasonable that all authentication mechanisms may take some time
  to intitialise, rather than just DNS.

[1] - https://www.digicert.com/subject-alternative-name.htm