AD-HOC fix (Renewal): Modify consumption of renewal variable to be int
When the work to initially implement the Lets Encrypt work was created,
the "renewal" variable was surrounded by `"` characters, a habit the
author picked up while working with yaml more generally.

However, it appears that unless it is of the type "int" (i.e. without
the `"` characters and numeric) it does not take effect. This appears to
mean that the default value is used, which is 10 days.

It is difficult to reproduce this issue. Originally when it presented,
this was presumed to be an issue in the generation of the fullchain
certificate, however further investigation shows the renewal works, and
seemed to work at the "8" day period. Ansible was run as part of a
deployment 2 days prior to expiry, and the default is 10[1]

[1] - https://github.com/ansible/ansible/blob/devel/lib/ansible/modules/web_infrastructure/letsencrypt.py#L858