AD-HOC fix (domain names) Change regex so it only prefies 'DNS:' once per Domain
While using this Role to setup a project, I got an error response from
lets encrypt api that was refusing the refusing the crt-request. A bit
of investigation shows that the alternative domain name was converted to
'DNS:example.comDNS:' The 'DNS:' at the end of the string should not be
there. The provided fix is making shure the regex is matching the whole
sting, so the alternate domain is now 'DNS:example.com'.