AdHoc (cert-creation) Enable recreation of certificates
It turns out that old certificates are not renewed or updated with new
configurations (for example additional or changed domains) while
delployments.

It turns out that removing the certifificate in the path
/etc/ssl/certs/XXXX.lets_encrypt.crt
is not touched even when it should be changed. Removing this file is
getting the renewing process into success. This indicates that the
let's encrypt component is not overwriting existing files.

This now fixed with writing the new certificate into an temporary
directory and copy it over to the cert dir later with ansible.