Adhoc (acme) Update acme to version 2
As letsencrypt has depricated acme_v1 and disabled it's acme_v1
endpoint. We need to change to acme_v2. This requires to use a new
interface. This change introduces an update, but this also requires an
update of ansible, that is reflected in the minimal supported version.

We do not need to support acme_v1 anymore, because service was
discuntinued.